+++ pages/zustand/index.tsx
@@ -1,7 +1,6 @@
-import { shaderMaterial } from '@react-three/drei'
 import * as THREE from 'three'
 import React, { Suspense, useRef, useState } from 'react'
-import { Canvas, useFrame, useThree, extend } from '@react-three/fiber'
+import { Canvas, useFrame, useThree } from '@react-three/fiber'
 import { Plane, useAspect, useTexture } from '@react-three/drei'
 import { EffectComposer, DepthOfField, Vignette } from '@react-three/postprocessing'
 import create from 'zustand'
@@ -11,6 +10,7 @@ import 'prismjs/components/prism-jsx.min'
 import Fireflies from '../../components/Fireflies'
 import Link from 'next/link'
 import SEO from 'components/Seo'
+import 'components/LayerMaterial'
 
 const useStore = create<{
   count: number
@@ -206,40 +206,3 @@ function Counter() {
   const count = useStore(state => state.count)
   return <h1>{count}</h1>  
 }`
-
-const LayerMaterial = shaderMaterial(
-  { textr: null, movementVector: [0, 0, 0], scaleFactor: 1, factor: 0, wiggle: 0, time: 0 },
-  ` uniform float time;
-    uniform vec2 resolution;
-    uniform float wiggle;
-    varying vec2 vUv;
-    varying vec3 vNormal;
-    void main()	{
-      vUv = uv;
-      vec3 transformed = vec3(position);
-      if (wiggle > 0.) {
-        float theta = sin(time + position.y) / 2.0 * wiggle;
-        float c = cos(theta);
-        float s = sin(theta);
-        mat3 m = mat3(c, 0, s, 0, 1, 0, -s, 0, c);
-        transformed = transformed * m;
-        vNormal = vNormal * m;
-      }      
-      gl_Position = projectionMatrix * modelViewMatrix * vec4(transformed, 1.);
-    }`,
-  ` uniform float time;
-    uniform vec2 resolution;
-    uniform float factor;
-    uniform float scaleFactor;
-    uniform vec3 movementVector;
-    uniform sampler2D textr;
-    varying vec2 vUv;
-    void main()	{
-      vec2 uv = vUv / scaleFactor + movementVector.xy * factor;
-      vec4 color = texture2D(textr, uv);
-      if (color.a < 0.1) discard;
-      gl_FragColor = vec4(color.rgb, .1);
-    }`
-)
-
-extend({ LayerMaterial })
